requirements.txt
absl-py==2.4.0
altair==6.0.0
asgiref==3.11.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
Django==6.0.3
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.60.1
fsspec==2025.9.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.78.0
gunicorn==25.1.0
h5py==3.15.1
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.4
moviepy==2.2.1
mpmath==1.3.0
namex==0.1.0
narwhals==2.17.0
networkx==3.5
numpy==2.2.6
opencv-contrib-python==4.13.0.92
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
polars==1.34.0
polars-runtime-32==1.34.0
proglog==0.1.12
protobuf==6.33.5
psutil==7.1.1
pyarrow==23.0.1
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.5
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.9.0
torchvision==0.24.0
tornado==6.5.4
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.218
ultralytics-thop==2.0.17
urllib3==2.5.0
watchdog==6.0.0
Werkzeug==3.1.6
wheel==0.46.3
whitenoise==6.12.0
wrapt==2.1.1
